NJTA-17-course-api

+++ src/main/java/net/njit/ms/cs/repository/CourseRepository.java
@@ -3,5 +3,5 @@ package net.njit.ms.cs.repository;
 import net.njit.ms.cs.model.entity.Course;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CourseRepository extends JpaRepository<Course, Long> {
+public interface CourseRepository extends JpaRepository<Course, Integer> {
 }
